fvMesh: Do not alter the moving state on construction
fvMesh no longer sets the mesh state to moving when meshPhi and old-time
volume data is present. This is unecessary because a solver should
always initialise a mesh update before any time-derivatives are
calculated. Setting the mesh as moving complicates the usage of some
utilities and post-processing applications in which time is not
incremented.

The meshPhi and old-time volume data are still read in if present to
ensure restart consistency.