reactingEulerFoam: wallDampingModels: Additional control parameters
Added two new user parameters to wallDampingModels which allow
additional control over damping to zero in the near vicinity of the
walls.

"zeroWallDist" is a distance from the walls below which the damping
function sets the value of the model to zero.

"zeroInNearWallCells" is a switch which sets the value of the model to
zero in wall adjacent cells, regardless of their actual distance from
the wall.

Patch contributed by Juho Peltola, VTT.