functionObjects: yPlus: Added support for phases
The yPlus function object can now take a "phase" keyword which defines
the name of the phase for which to write yPlus. For example, to write
yPlus for a phase named "liquid", the following entry can be used:

    yPlus1
    {
        type        yPlus;
        libs        ("libfieldFunctionObjects.so");
        phase       liquid;
    }

Note that this will only be necessary in Euler-Euler type simulations
where the phases have separate turbulence models. For VoF, the phase
name will not be required.

Patch contributed by Juho Peltola, VTT.