reactions: Added surfaceArrheniusReactionRate
This is a modified Arrhenius reaction rate where the reaction rate is
multiplied by a surface area per unit volume.

The reaction rate is given by:

    k = (A * T^beta * exp(-Ta/T))*a

Where a is the surface area per unit volume, the name of which is
specified by the user. This enables, for example, the reaction rate to
depend on the surface area concentration of catalyst particles.

Patch contributed by VTT Technical Research Centre of Finland Ltd and
Institute of Fluid Dynamics, Helmholtz-Zentrum Dresden - Rossendorf (HZDR).